+++ CotEditor/Sources/CustomSurroundStringViewController.swift
@@ -8,7 +8,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2017-2018 1024jp
+//  © 2017-2019 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -29,8 +29,8 @@ final class CustomSurroundStringViewController: NSViewController {
     
     // MARK: Private Properties
     
-    @objc private dynamic var beginString: String = ""
-    @objc private dynamic var endString: String = ""
+    @objc dynamic var beginString: String = ""
+    @objc dynamic var endString: String = ""
     
     @IBOutlet private weak var beginStringField: NSTextField?
     @IBOutlet private weak var endStringField: NSTextField?