+++ CotEditor/Sources/EditorTextView.swift
@@ -68,6 +68,8 @@ final class EditorTextView: NSTextView, Themable, CurrentLineHighlighting, Multi
     var initialMagnificationScale: CGFloat = 0
     var deferredMagnification: CGFloat = 0
     
+    private(set) lazy var customSurroundStringViewController = self.instantinateSurroundStringViewController()
+    
     
     // MARK: Private Properties